Implemented (dummy) workaround to support the cursor on the VGA adapter

+++ x8086NetEmu/Adapters/Adapter.vb
@@ -47,7 +47,7 @@
     End Property
 
     Public MustOverride Overrides ReadOnly Property Description As String
-    Public MustOverride Overrides Function [In](port As UInt32) As UInt32
-    Public MustOverride Overrides Sub Out(port As UInt32, value As UInt32)
+    Public MustOverride Overrides Function [In](port As UInt32) As UInt16
+    Public MustOverride Overrides Sub Out(port As UInt32, value As UInt16)
     Public MustOverride Overrides ReadOnly Property Name As String
 End Class
