+++ x8086NetEmu/Chipset/PIC8259.vb
@@ -117,7 +117,7 @@
         If master IsNot Nothing Then UpdateSlaveOutput()
     End Sub
 
-    Public Overrides Function [In](port As UInt32) As UInt32
+    Public Overrides Function [In](port As UInt32) As UInt16
         If (port And 1) = 0 Then
             ' A0 == 0
             If pollMode Then
@@ -131,7 +131,7 @@
         End If
     End Function
 
-    Public Overrides Sub Out(port As UInt32, value As UInt32)
+    Public Overrides Sub Out(port As UInt32, value As UInt16)
         If (port And 1) = 0 Then
             ' A0 = 0
             If (value And &H10) <> 0 Then
